Consolidated some packages
We had a few packages with only a single file that could logically fit
into other packages better. Some of these such as the repository package
could be split into subpackages in the future (e.g., if we had more than
just a file-based implementation).
