+++ ReferenceAppKotlin/app/src/main/java/com/android/tv/reference/auth/NoFirebaseFragment.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.android.tv.reference.welcome
+package com.android.tv.reference.auth
 
 import android.os.Bundle
 import android.view.LayoutInflater
