+++ ReferenceAppKotlin/app/src/main/java/com/android/tv/reference/playback/PlaybackStateMachine.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.android.tv.reference.shared.playback
+package com.android.tv.reference.playback
 
 import com.android.tv.reference.shared.datamodel.Video
