+++ ReferenceAppKotlin/app/src/main/java/com/android/tv/reference/shared/watchprogress/WatchProgressPlaybackStateListener.kt
@@ -16,7 +16,7 @@
 package com.android.tv.reference.shared.watchprogress
 
 import com.android.tv.reference.playback.PlaybackStateListener
-import com.android.tv.reference.shared.playback.VideoPlaybackState
+import com.android.tv.reference.playback.VideoPlaybackState
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.launch
