+++ ReferenceAppKotlin/app/src/test/java/com/android/tv/reference/shared/watchprogress/WatchProgressPlaybackStateListenerTest.kt
@@ -24,7 +24,7 @@ import com.android.tv.reference.coroutines.CoroutinesTestRule
 import com.android.tv.reference.lifecycle.ext.getOrAwaitValue
 import com.android.tv.reference.shared.datamodel.Video
 import com.android.tv.reference.shared.datamodel.VideoType
-import com.android.tv.reference.shared.playback.VideoPlaybackState
+import com.android.tv.reference.playback.VideoPlaybackState
 import com.google.common.truth.Truth.assertThat
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.test.TestCoroutineScope
